Realm-AI/requirements.txt
streamlit
python-dotenv
langchain
transformers
sentence-transformers
torch
faiss-cpu
numpy
langchain_community
pypdf
tiktoken
flask